requirements.txt
# FitScore - CrossFit Competition Management System
# Python dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.13.0

# Authentication
passlib[bcrypt]>=1.7.4
bcrypt==4.2.0
python-jose[cryptography]>=3.3.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Templates
jinja2>=3.1.0

# Export (PDF, Excel)
openpyxl>=3.1.0
reportlab>=4.0.0

# Development
python-dotenv>=1.0.0

# Production (optional)
gunicorn>=21.0.0
